% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_longitudinal_data.R
\name{.generate_dummy_service_sector_data}
\alias{.generate_dummy_service_sector_data}
\title{.generate_dummy_service_sector_data}
\usage{
.generate_dummy_service_sector_data(
  par_parameters = list(n_patients = 30, n_patients_offset = 0, seed = 13),
  service_sector_data_version = "R10v2"
)
}
\arguments{
\item{par_parameters}{PARAM_DESCRIPTION, Default: list(n_patients = 30, n_patients_offset = 0, seed = 13)}

\item{service_sector_data_version}{PARAM_DESCRIPTION, Default: 'R10v2'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Create a tibble with dummy data in the service sector format.
This function is ran in parallel by function generate_dummy_service_sector_data.
}
