% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_hla_imputed.R
\name{generate_dummy_hla_imputed_data}
\alias{generate_dummy_hla_imputed_data}
\title{generate_dummy_hla_imputed}
\usage{
generate_dummy_hla_imputed_data(
  hla_imputed_level_data_version = "R12v1",
  n_patients_minimum = 100,
  seed = 13
)
}
\arguments{
\item{n_patients_minimum}{PARAM_DESCRIPTION, Default: 100}

\item{seed}{PARAM_DESCRIPTION, Default: 13}

\item{vision_level_data_version}{PARAM_DESCRIPTION, Default: 'R12v1'}

\item{description}{PARAM_DESCRIPTION, Default: HLA imputed dummy data generator}
}
\value{
list of created tables
}
\description{
FUNCTION_DESCRIPTION
}
