% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_longitudinal_data.R
\name{generate_dummy_longitudinal_data}
\alias{generate_dummy_longitudinal_data}
\title{Title}
\usage{
generate_dummy_longitudinal_data(
  output_folder,
  longitudinal_data_version = "DF6v2",
  n_patients = 100,
  seed = 13,
  nTreaths = 3
)
}
\arguments{
\item{nTreaths}{}
}
\value{

}
\description{
Title
}
