% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_minimum_extended_data.R
\name{generate_dummy_minimum_extended_data}
\alias{generate_dummy_minimum_extended_data}
\title{generate_dummy_minimum_extended_data}
\usage{
generate_dummy_minimum_extended_data(
  person_level_data_version = "R10v1",
  n_patients_minimum = 100,
  seed = 13,
  service_sector_data = NULL
)
}
\arguments{
\item{person_level_data_version}{PARAM_DESCRIPTION, Default: 'R10v1'}

\item{n_patients_minimum}{PARAM_DESCRIPTION, Default: 100}

\item{seed}{PARAM_DESCRIPTION, Default: 13}

\item{service_sector_data}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
list of created tables
}
\description{
FUNCTION_DESCRIPTION
}
