% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_mothers_birth_register.R
\name{generate_dummy_mothers_birth_register_data}
\alias{generate_dummy_mothers_birth_register_data}
\title{generate_dummy_birth_mother_data}
\usage{
generate_dummy_mothers_birth_register_data(
  birth_mother_level_data_version = "R12v1",
  n_patients_minimum = 100,
  seed = 13,
  service_sector_data = NULL
)
}
\arguments{
\item{birth_mother_level_data_version}{PARAM_DESCRIPTION, Default: 'R12v3'}

\item{n_patients_minimum}{PARAM_DESCRIPTION, Default: 100}

\item{seed}{PARAM_DESCRIPTION, Default: 13}

\item{service_sector_data}{PARAM_DESCRIPTION, Default: NULL}

\item{description}{PARAM_DESCRIPTION, Default: Birth Mother dummy data generator}
}
\value{
list of created tables
}
\description{
FUNCTION_DESCRIPTION
}
