% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_other_drugs_data.R
\name{generate_dummy_other_drugs_data}
\alias{generate_dummy_other_drugs_data}
\title{generate_dummy_other_drugs_data}
\usage{
generate_dummy_other_drugs_data(
  other_drugs_level_data_version = "R12v1",
  n_patients_minimum = 100,
  seed = 13
)
}
\arguments{
\item{other_drugs_level_data_version}{PARAM_DESCRIPTION, Default: 'R12v1'}

\item{n_patients_minimum}{PARAM_DESCRIPTION, Default: 100}

\item{seed}{PARAM_DESCRIPTION, Default: 13}

\item{description}{PARAM_DESCRIPTION, Default: Vision dummy data generator}
}
\value{
list of created tables
}
\description{
FUNCTION_DESCRIPTION
}
