% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_spirometry_register.R
\name{generate_dummy_spirometry_data}
\alias{generate_dummy_spirometry_data}
\title{generate_dummy_spirometry_data}
\usage{
generate_dummy_spirometry_data(
  spirometry_level_data_version = "R12v1",
  n_patients_minimum = 100,
  seed = 13,
  minimum_extended = NULL
)
}
\arguments{
\item{spirometry_level_data_version}{PARAM_DESCRIPTION, Default: 'R12v1'}

\item{n_patients_minimum}{PARAM_DESCRIPTION, Default: 100}

\item{seed}{PARAM_DESCRIPTION, Default: 13}

\item{minimum_extended}{PARAM_DESCRIPTION, Default: NULL}

\item{description}{PARAM_DESCRIPTION, Default: spirometry dummy data generator}
}
\value{
list of created tables
}
\description{
FUNCTION_DESCRIPTION
}
