% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_vaccination_register.R
\name{generate_dummy_vaccination_data}
\alias{generate_dummy_vaccination_data}
\title{generate_dummy_vaccination_data}
\usage{
generate_dummy_vaccination_data(
  vaccination_level_data_version = "R12v1",
  n_patients_minimum = 100,
  seed = 13,
  service_sector_data = NULL,
  kanta_medication_delivery_data = NULL
)
}
\arguments{
\item{vaccination_level_data_version}{PARAM_DESCRIPTION, Default: 'R12v1'}

\item{n_patients_minimum}{PARAM_DESCRIPTION, Default: 100}

\item{seed}{PARAM_DESCRIPTION, Default: 13}

\item{service_sector_data}{PARAM_DESCRIPTION, Default: NULL}

\item{kanta_medication_delivery_data}{PARAM_DESCRIPTION, Default: NULL}

\item{description}{PARAM_DESCRIPTION, Default: vaccination dummy data generator}
}
\value{
list of created tables
}
\description{
FUNCTION_DESCRIPTION
}
