% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_vision_register.R
\name{generate_dummy_vision_register_data}
\alias{generate_dummy_vision_register_data}
\title{generate_dummy_vision_data}
\usage{
generate_dummy_vision_register_data(
  vision_level_data_version = "R12v1",
  n_patients_minimum = 100,
  seed = 13,
  minimum_extended = NULL
)
}
\arguments{
\item{vision_level_data_version}{PARAM_DESCRIPTION, Default: 'R12v4'}

\item{n_patients_minimum}{PARAM_DESCRIPTION, Default: 100}

\item{seed}{PARAM_DESCRIPTION, Default: 13}

\item{minimum_extended}{PARAM_DESCRIPTION, Default: NULL}

\item{description}{PARAM_DESCRIPTION, Default: Vision dummy data generator}
}
\value{
list of created tables
}
\description{
FUNCTION_DESCRIPTION
}
