% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanReportToTibble.R
\name{scanReportToTibble}
\alias{scanReportToTibble}
\title{Convert Scan Report to Tibble}
\usage{
scanReportToTibble(scan_report, n_rows)
}
\arguments{
\item{scan_report}{The scan report data frame.}

\item{n_rows}{Number of rows for the resulting tibble.}
}
\description{
Converts a scan report from WaitRabit to a tibble with specified number of rows.
}
