% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summary_data_versions_list}
\alias{summary_data_versions_list}
\title{Collection of summary tables for building the dummy datasets.}
\format{
An object of class \code{list} of length 1.
}
\usage{
summary_data_versions_list
}
\description{
Collection of summary tables for building the dummy datasets.
}
\keyword{datasets}
