% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_all_datasets_to_files.R
\name{generate_all_dummy_data_to_files}
\alias{generate_all_dummy_data_to_files}
\title{Generate All Dummy Data to Files}
\usage{
generate_all_dummy_data_to_files(
  output_folder,
  service_sector_data_version = "R10v2",
  person_level_data_version = "R10v1",
  n_patients_minimum = 100,
  per_patients_service_sector = 0.99,
  seed = 13,
  nTreaths = (parallel::detectCores() - 2)
)
}
\arguments{
\item{output_folder}{Path to the folder where the generated data will be saved.}

\item{service_sector_data_version}{Version of the service sector data (default: "R10v2").}

\item{person_level_data_version}{Version of the person level data (default: "R10v1").}

\item{n_patients_minimum}{Number of patients to generate (default: 100).}

\item{per_patients_service_sector}{Proportion of patients relative to n_patients_minimum to generate service sector data from  (default: 0.99).}

\item{seed}{Seed value for random number generation (default: 13).}

\item{nTreaths}{Number of threads for parallel processing (default: parallel::detectCores() - 2).}
}
\description{
Generates dummy data for service sector and person level data in the
"summary_data_versions_list" internal variable, and saves them to files.
}
