% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_longitudinal_data.R
\name{generate_dummy_longitudinal_data}
\alias{generate_dummy_longitudinal_data}
\title{generate_dummy_longitudinal_data}
\usage{
generate_dummy_longitudinal_data(
  output_folder,
  longitudinal_data_version = "R10v1",
  n_patients = 30,
  n_cuts = 3,
  seed = 13,
  nTreaths = 2
)
}
\arguments{
\item{output_folder}{directory where to output the generated data}

\item{longitudinal_data_version}{at the moment only DF6v2 is avalilable}

\item{n_patients}{number of random patients to genenrate}

\item{seed}{seed used in the random processes}

\item{nTreaths}{number of cores to use when using paraller generation (if large you can use parallel::detectCores() -1)}
}
\description{
Produces longituinal and baseline data in folder \code{output_folder} for \code{n_patients} number of patietns
}
