% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_service_sector_data.R
\name{generate_dummy_service_sector_data}
\alias{generate_dummy_service_sector_data}
\title{generate_dummy_service_sector_data}
\usage{
generate_dummy_service_sector_data(
  service_sector_data_version = "R12v1",
  n_patients = 30,
  n_cuts = nTreaths,
  seed = 13,
  nTreaths = (parallel::detectCores() - 2)
)
}
\arguments{
\item{service_sector_data_version}{at the moment only DF6v2 is available, Default: 'R12v2'}

\item{n_patients}{number of random patients to generate, Default: 30}

\item{n_cuts}{PARAM_DESCRIPTION, Default: 3}

\item{seed}{seed used in the random processes, Default: 13}

\item{nTreaths}{number of cores to use when using parallel generation, Default: (parallel::detectCores() -1)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Generates a file with dummy data in the service sector format.
}
